requirements.txt
torch>=1.9.0
torchvision
numpy
pillow
scikit-image
pyyaml